Move confighandle.functions package to config package.
