+++ src/main/java/org/mvplugins/multiverse/core/command/MVCommandCompletions.java
@@ -31,7 +31,7 @@ import org.jvnet.hk2.annotations.Service;
 
 import org.mvplugins.multiverse.core.anchor.AnchorManager;
 import org.mvplugins.multiverse.core.config.CoreConfig;
-import org.mvplugins.multiverse.core.confighandle.functions.DefaultSuggesterProvider;
+import org.mvplugins.multiverse.core.config.functions.DefaultSuggesterProvider;
 import org.mvplugins.multiverse.core.config.handle.PropertyModifyAction;
 import org.mvplugins.multiverse.core.destination.DestinationInstance;
 import org.mvplugins.multiverse.core.destination.DestinationsProvider;
