+++ src/main/java/org/mvplugins/multiverse/core/config/functions/DefaultSerializerProvider.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.confighandle.functions;
+package org.mvplugins.multiverse.core.config.functions;
 
 import java.util.HashMap;
 import java.util.Locale;
