--- src/main/java/org/mvplugins/multiverse/core/confighandle/functions/package-info.java
@@ -1,1 +1,0 @@
-package org.mvplugins.multiverse.core.confighandle.functions;
