+++ src/main/java/org/mvplugins/multiverse/core/confighandle/node/ListConfigNode.java
@@ -16,7 +16,7 @@ import io.vavr.control.Try;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import org.mvplugins.multiverse.core.confighandle.functions.*;
+import org.mvplugins.multiverse.core.config.functions.*;
 
 /**
  * A config node that contains a list of values.
